+++ pyccel/codegen/make_pipeline.py
@@ -28,7 +28,7 @@ from pyccel.errors.errors import (
 )
 from pyccel.naming import name_clash_checkers
 from pyccel.parser.parser import Parser
-from pyccel.parser.scope import Scope
+from pyccel.plugins.plugin_tools import get_build_generation_class
 from pyccel.utilities.stage import PyccelStage
 
 from .compiling.compilers import Compiler, get_condaless_search_path
@@ -58,6 +58,7 @@ def execute_pyccel_make(
     accelerators,
     conda_warnings,
     build_code,
+    plugin_manager,
 ):
     """
     Run `pyccel make` on the provided files.
@@ -105,6 +106,8 @@ def execute_pyccel_make(
         Specify the level of Conda warnings to display (choices: off, basic, verbose).
     build_code : bool
         Indicates if the build commands should be run.
+    plugin_manager : pluggy.PluginManager
+        The plugin manager used to connect activated plugins.
     """
     start = time.time()
     timers = {}
@@ -144,9 +147,6 @@ def execute_pyccel_make(
     Compiler.acceptable_bin_paths = get_condaless_search_path(conda_warnings)
     compiler = Compiler(compiler_family, debug)
 
-    # Get compiler object
-    Scope.name_clash_checker = name_clash_checkers[language]
-
     start_syntax = time.time()
     timers["Initialisation"] = start_syntax - start
 
@@ -160,7 +160,15 @@ def execute_pyccel_make(
     cwd = os.getcwd()
     files = [f.relative_to(cwd) if f.is_absolute() else f for f in files]
 
-    parsers = {f: Parser(f.absolute(), output_folder=folder) for f in files}
+    parsers = {
+        f: Parser(
+            f.absolute(),
+            output_folder=folder,
+            name_clash_checker=name_clash_checkers[language],
+            plugin_manager=plugin_manager,
+        )
+        for f in files
+    }
 
     to_remove = []
     for f, p in parsers.items():
@@ -207,7 +215,9 @@ def execute_pyccel_make(
         semantic_parser = p.semantic_parser
         start_codegen = time.time()
         # Generate low-level code file
-        codegen = Codegen(semantic_parser, f, language, verbose)
+        codegen = Codegen(
+            semantic_parser, f, language, verbose, plugin_manager=plugin_manager
+        )
         fname = (pyccel_dirpath / f).with_suffix("")
         output_dir = fname.parent
         os.makedirs(output_dir, exist_ok=True)
@@ -227,7 +237,9 @@ def execute_pyccel_make(
             shutil.copyfile(fname, new_location)
         else:
             start_wrapper_creation = time.time()
-            wrappergen = Wrappergen(codegen, codegen.name, language, verbose)
+            wrappergen = Wrappergen(
+                codegen, codegen.name, language, verbose, plugin_manager=plugin_manager
+            )
             wrappergen.wrap(str((base_dirpath / f).parent))
             timers["Wrapper creation"] += time.time() - start_wrapper_creation
 
@@ -290,7 +302,10 @@ def execute_pyccel_make(
         base_dirpath, targets.values(), printed_languages, stdlib_deps
     )
 
-    build_sys = build_system_handler[build_system](
+    BuildGenClass = get_build_generation_class(
+        plugin_manager, build_system_handler.get(build_system, None), build_system
+    )
+    build_sys = BuildGenClass(
         pyccel_dirpath,
         base_dirpath,
         folder,