+++ pyccel/codegen/wrappergen.py
@@ -9,10 +9,10 @@ A module containing the Wrappergen class which is responsible for the generation
 
 from pathlib import Path
 
+from pyccel.plugins.plugin_tools import get_wrapper_class, get_wrapper_codegen_class
+
 from ..ast.core import ModuleHeader
 from ..errors.errors import Errors
-from ..naming import name_clash_checkers
-from ..parser.scope import Scope
 from ..utilities.stage import PyccelStage
 from .codegen import _extension_registry, _header_extension_registry
 from .printing.cwrappercode import CWrapperCodePrinter
@@ -23,16 +23,15 @@ from .wrapper.cpp_to_python_wrapper import CppToPythonWrapper
 from .wrapper.fortran_to_c_wrapper import FortranToCWrapper
 
 wrapper_registry = {
-    "fortran": [FortranToCWrapper, CToPythonWrapper],
-    "c": [CToPythonWrapper],
-    "c++": [CppToPythonWrapper],
-    "python": [],
+    "fortran": (FortranToCWrapper, "c"),
+    "c": (CToPythonWrapper, "python"),
+    "c++": (CppToPythonWrapper, "python"),
 }
 
 printer_registry = {
-    FortranToCWrapper: FCodePrinter,
-    CToPythonWrapper: CWrapperCodePrinter,
-    CppToPythonWrapper: PyBindCodePrinter,
+    "fortran": FCodePrinter,
+    "c": CWrapperCodePrinter,
+    "c++": PyBindCodePrinter,
 }
 
 pyccel_stage = PyccelStage()
@@ -57,9 +56,11 @@ class Wrappergen:
         The language which the printer should print to.
     verbose : int
         The level of verbosity.
+    plugin_manager : pluggy.PluginManager
+        The plugin manager used to connect activated plugins.
     """
 
-    def __init__(self, codegen, name, language, verbose):
+    def __init__(self, codegen, name, language, verbose, plugin_manager):
         pyccel_stage.set_stage("cwrapper")
         self._ast = codegen.ast
         self._name = name
@@ -67,8 +68,21 @@ class Wrappergen:
         self._verbose = verbose
         self._wrapper_ast = []
 
-        self._wrapper_types = wrapper_registry[language]
-        self._printer_types = [printer_registry[w] for w in self._wrapper_types]
+        languages = []
+        self._wrapper_types = []
+        while language != "python":
+            start_language = language
+            languages.append(start_language)
+            wrapper_base_class, language = wrapper_registry.get(language, (None, None))
+            wrapper, language = get_wrapper_class(
+                plugin_manager, wrapper_base_class, start_language, language
+            )
+            self._wrapper_types.append(wrapper)
+
+        self._printer_types = [
+            get_wrapper_codegen_class(plugin_manager, printer_registry.get(l, None), l)
+            for l in languages
+        ]
         self._additional_imports = [{} for _ in self._wrapper_types]
 
     def wrap(self, sharedlib_dirpath):
@@ -83,7 +97,6 @@ class Wrappergen:
         sharedlib_dirpath : str
             The folder where the generated .so file will be located.
         """
-        current_name_clash_checker = Scope.name_clash_checker
         ast = self._ast
         for Wrapper in self._wrapper_types:
             if self._verbose:
@@ -92,9 +105,6 @@ class Wrappergen:
                     self._name,
                 )
 
-            Scope.name_clash_checker = name_clash_checkers[
-                Wrapper.start_language.lower()
-            ]
             wrapper = Wrapper(sharedlib_dirpath, verbose=self._verbose)
 
             ast = wrapper.wrap(ast)
@@ -102,8 +112,6 @@ class Wrappergen:
 
             if errors.has_errors():
                 break
-
-        Scope.name_clash_checker = current_name_clash_checker
 
     def print(self, dirpath):
         """