+++ tests/ast/test_basic.py
@@ -1,6 +1,7 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 import os
 
+import pluggy
 import pytest
 
 from pyccel.ast.basic import PyccelAstNode
@@ -10,6 +11,7 @@ from pyccel.ast.literals import LiteralInteger
 from pyccel.ast.operators import PyccelAdd, PyccelMinus, PyccelMul, PyccelOperator
 from pyccel.ast.variable import Variable
 from pyccel.errors.errors import Errors
+from pyccel.naming import name_clash_checkers
 from pyccel.parser.parser import Parser
 
 base_dir = os.path.dirname(os.path.realpath(__file__))
@@ -17,7 +19,14 @@ path_dir = os.path.join(base_dir, "scripts")
 
 
 def get_functions(filename):
-    pyccel = Parser(filename, output_folder=os.getcwd())
+    plugin_manager = pluggy.PluginManager("pyccel")
+
+    pyccel = Parser(
+        filename,
+        output_folder=os.getcwd(),
+        name_clash_checker=name_clash_checkers["python"],
+        plugin_manager=plugin_manager,
+    )
     errors = Errors()
 
     ast = pyccel.parse(verbose=0)