+++ tests/syntax/test_syntax.py
@@ -3,9 +3,11 @@
 
 import os
 
+import pluggy
 import pytest
 
 from pyccel.errors.errors import Errors
+from pyccel.naming import name_clash_checkers
 from pyccel.parser.parser import Parser
 
 base_dir = os.path.dirname(os.path.realpath(__file__))
@@ -22,7 +24,14 @@ def test_syntax(f):
     errors = Errors()
     errors.reset()
 
-    pyccel = Parser(f, output_folder=os.getcwd())
+    plugin_manager = pluggy.PluginManager("pyccel")
+
+    pyccel = Parser(
+        f,
+        output_folder=os.getcwd(),
+        name_clash_checker=name_clash_checkers["python"],
+        plugin_manager=plugin_manager,
+    )
     pyccel.parse(verbose=0)
 
     # Assert syntactic success